requirements.txt
asgiref==3.2.7
Django==3.0.8
django-ckeditor==5.9.0
django-extensions==2.2.9
django-js-asset==1.2.2
django-mptt==0.11.0
gunicorn==20.0.4
Pillow==7.1.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
whitenoise==5.1.0


gunicorn==20.0.4
Brotli == 1.0.7


djangorestframework==3.11.0
django-filter==2.3.0